versions/v1/requirements.txt
feedparser
GitPython